AllPhotos unit test done
